Change package of jetty-mongo for consistency.
