+++ jetty-session/jetty-mongodb/src/main/config/etc/sessions/mongo/session-store-by-uri.xml
@@ -9,7 +9,7 @@
   <!-- ===================================================================== -->
   <Call name="addBean">
    <Arg>
-    <New id="sessionDataStoreFactory" class="org.eclipse.jetty.mongodb.MongoSessionDataStoreFactory">      
+    <New id="sessionDataStoreFactory" class="org.eclipse.jetty.session.mongodb.MongoSessionDataStoreFactory">      
        <Set name="dbName"><Property name="jetty.session.mongo.dbName" default="HttpSessions" /></Set>
        <Set name="collectionName"><Property name="jetty.session.mongo.collectionName" default="jettySessions" /></Set>
        <Set name="gracePeriodSec"><Property name="jetty.session.gracePeriod.seconds" default="3600" /></Set>
