+++ jetty-session/jetty-mongodb/src/main/java/org/eclipse/jetty/session/mongodb/MongoSessionDataStore.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.mongodb;
+package org.eclipse.jetty.session.mongodb;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
