+++ tests/test-sessions/test-mongodb-sessions/src/test/java/org/eclipse/jetty/session/mongodb/ClusteredSessionScavengingTest.java
@@ -11,16 +11,16 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.mongodb;
+package org.eclipse.jetty.session.mongodb;
 
-import org.eclipse.jetty.server.session.AbstractClusteredInvalidationSessionTest;
+import org.eclipse.jetty.server.session.AbstractClusteredSessionScavengingTest;
 import org.eclipse.jetty.server.session.SessionDataStoreFactory;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Testcontainers(disabledWithoutDocker = true)
-public class ClusteredInvalidateSessionTest extends AbstractClusteredInvalidationSessionTest
+public class ClusteredSessionScavengingTest extends AbstractClusteredSessionScavengingTest
 {
     @BeforeAll
     public static void beforeClass() throws Exception
